# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface FileNoteCoverageItem {
  label: string;
  value: string;
  detail?: string;
}

export interface FileNoteRewriteInput {
  /** @minLength 1 */
  contact: string;
  meetingType?: string;
  date?: string;
  /** @minLength 1 */
  note: string;
  /** Confirmed discussion topics; when present the note is enhanced rather than rewritten from scratch */
  coverage?: FileNoteCoverageItem[];
}

export interface FileNoteRewriteResult {
  note: string;
}

/**
 * Optional pre-meeting briefing context (subset of the stored briefing)
 */
export interface SourceOfWealthBriefingInput {
  summary?: string;
  talkingPoints?: string[];
  referralRoutes?: string[];
  recommendedApproach?: string;
}

export interface SourceOfWealthDraftInput {
  /** @minLength 1 */
  clientName: string;
  /** The banker's meeting note (may be empty when only a briefing is available) */
  note?: string;
  /** Optional confirmed discussion topics to ground the draft */
  coverage?: FileNoteCoverageItem[];
  briefing?: SourceOfWealthBriefingInput;
}

export interface SourceOfWealthStatement {
  overview: string;
  employment: string;
  compensation: string;
  assetGrowth: string;
  wealthEvents: string;
  plausibility: string;
}

export interface SourceOfWealthDraftResult {
  statement: SourceOfWealthStatement;
}

export type ReviewType = typeof ReviewType[keyof typeof ReviewType];


export const ReviewType = {
  onboarding: 'onboarding',
  periodic_review: 'periodic_review',
  trigger_event: 'trigger_event',
} as const;

export type RiskRating = typeof RiskRating[keyof typeof RiskRating];


export const RiskRating = {
  standard: 'standard',
  enhanced: 'enhanced',
} as const;

export type AssessmentStatus = typeof AssessmentStatus[keyof typeof AssessmentStatus];


export const AssessmentStatus = {
  draft: 'draft',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

/**
 * Free-form record of every captured response, keyed by field id
 */
export type AssessmentData = { [key: string]: unknown };

export interface Assessment {
  id: number;
  clientName: string;
  /** @nullable */
  clientReference?: string | null;
  /** @nullable */
  relationshipManager?: string | null;
  reviewType?: ReviewType;
  riskRating?: RiskRating;
  status: AssessmentStatus;
  /** Free-form record of every captured response, keyed by field id */
  data: AssessmentData;
  createdAt: string;
  updatedAt: string;
}

export interface AssessmentSummary {
  id: number;
  clientName: string;
  /** @nullable */
  clientReference?: string | null;
  /** @nullable */
  relationshipManager?: string | null;
  reviewType?: ReviewType;
  riskRating?: RiskRating;
  status: AssessmentStatus;
  createdAt: string;
  updatedAt: string;
}

export type AssessmentInputData = { [key: string]: unknown };

export interface AssessmentInput {
  /** @minLength 1 */
  clientName: string;
  /** @nullable */
  clientReference?: string | null;
  /** @nullable */
  relationshipManager?: string | null;
  reviewType?: ReviewType;
  riskRating?: RiskRating;
  status?: AssessmentStatus;
  data?: AssessmentInputData;
}

export type AssessmentUpdateData = { [key: string]: unknown };

export interface AssessmentUpdate {
  /** @minLength 1 */
  clientName?: string;
  /** @nullable */
  clientReference?: string | null;
  /** @nullable */
  relationshipManager?: string | null;
  reviewType?: ReviewType;
  riskRating?: RiskRating;
  status?: AssessmentStatus;
  data?: AssessmentUpdateData;
}

export interface StatusCount {
  status: AssessmentStatus;
  count: number;
}

export interface RiskCount {
  riskRating: RiskRating;
  count: number;
}

export interface Overview {
  total: number;
  byStatus: StatusCount[];
  byRisk: RiskCount[];
  recentlyUpdated: AssessmentSummary[];
}

export type ProspectStatus = typeof ProspectStatus[keyof typeof ProspectStatus];


export const ProspectStatus = {
  identified: 'identified',
  researching: 'researching',
  briefed: 'briefed',
  outreach: 'outreach',
  converted: 'converted',
  dormant: 'dormant',
} as const;

export interface BriefingSource {
  title: string;
  url: string;
}

export interface Briefing {
  summary: string;
  talkingPoints: string[];
  referralRoutes: string[];
  recommendedApproach: string;
  sources: BriefingSource[];
  generatedAt: string;
}

/**
 * Free-form profile, channels and operational-question notes, keyed by catalog id
 */
export type ProspectData = { [key: string]: unknown };

export interface Prospect {
  id: number;
  name: string;
  /** @nullable */
  segment?: string | null;
  /** @nullable */
  relationshipManager?: string | null;
  status: ProspectStatus;
  /** Free-form profile, channels and operational-question notes, keyed by catalog id */
  data: ProspectData;
  briefing?: Briefing | null;
  /** @nullable */
  convertedAssessmentId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProspectSummary {
  id: number;
  name: string;
  /** @nullable */
  segment?: string | null;
  /** @nullable */
  relationshipManager?: string | null;
  status: ProspectStatus;
  hasBriefing: boolean;
  /** @nullable */
  convertedAssessmentId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type ProspectInputData = { [key: string]: unknown };

export interface ProspectInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  segment?: string | null;
  /** @nullable */
  relationshipManager?: string | null;
  status?: ProspectStatus;
  data?: ProspectInputData;
}

export type ProspectUpdateData = { [key: string]: unknown };

export interface ProspectUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  segment?: string | null;
  /** @nullable */
  relationshipManager?: string | null;
  status?: ProspectStatus;
  data?: ProspectUpdateData;
}