# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /file-notes/rewrite:
    post:
      operationId: rewriteFileNote
      tags: [file-notes]
      summary: Rewrite or enhance a meeting file note into professional private-banking format
      description: >-
        Stateless AI helper used by both the prospect and client file-note
        panels. Without `coverage` it rewrites the raw note into a structured
        professional file note; with `coverage` it enhances an existing draft by
        weaving in the banker's confirmed discussion topics.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileNoteRewriteInput"
      responses:
        "200":
          description: The rewritten or enhanced file note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileNoteRewriteResult"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Rewrite generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /source-of-wealth/draft:
    post:
      operationId: draftSourceOfWealth
      tags: [source-of-wealth]
      summary: Draft a Source of Wealth statement from a meeting note and briefing
      description: >-
        Stateless AI helper for the client assessment. Synthesises a
        regulator-facing Source of Wealth statement from the banker's meeting
        note, optional confirmed discussion topics, and an optional pre-meeting
        briefing. Returns six narrative sections; any section with no basis in
        the inputs is returned empty for the banker to complete. Never
        fabricates facts; plausibility inferences are framed as such.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceOfWealthDraftInput"
      responses:
        "200":
          description: The drafted Source of Wealth statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceOfWealthDraftResult"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Draft generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    FileNoteCoverageItem:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: string
        detail:
          type: string
    FileNoteRewriteInput:
      type: object
      required: [contact, note]
      properties:
        contact:
          type: string
          minLength: 1
        meetingType:
          type: string
        date:
          type: string
        note:
          type: string
          minLength: 1
        coverage:
          type: array
          description: Confirmed discussion topics; when present the note is enhanced rather than rewritten from scratch
          items:
            $ref: "#/components/schemas/FileNoteCoverageItem"
    FileNoteRewriteResult:
      type: object
      required: [note]
      properties:
        note:
          type: string
    SourceOfWealthBriefingInput:
      type: object
      description: Optional pre-meeting briefing context (subset of the stored briefing)
      properties:
        summary:
          type: string
        talkingPoints:
          type: array
          items:
            type: string
        referralRoutes:
          type: array
          items:
            type: string
        recommendedApproach:
          type: string
    SourceOfWealthDraftInput:
      type: object
      required: [clientName]
      properties:
        clientName:
          type: string
          minLength: 1
        note:
          type: string
          description: The banker's meeting note (may be empty when only a briefing is available)
        coverage:
          type: array
          description: Optional confirmed discussion topics to ground the draft
          items:
            $ref: "#/components/schemas/FileNoteCoverageItem"
        briefing:
          $ref: "#/components/schemas/SourceOfWealthBriefingInput"
    SourceOfWealthStatement:
      type: object
      required: [overview, employment, compensation, assetGrowth, wealthEvents, plausibility]
      properties:
        overview:
          type: string
        employment:
          type: string
        compensation:
          type: string
        assetGrowth:
          type: string
        wealthEvents:
          type: string
        plausibility:
          type: string
    SourceOfWealthDraftResult:
      type: object
      required: [statement]
      properties:
        statement:
          $ref: "#/components/schemas/SourceOfWealthStatement"
...