# lib/api-zod/src/generated/types/sourceOfWealthBriefingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Optional pre-meeting briefing context (subset of the stored briefing)
 */
export interface SourceOfWealthBriefingInput {
  summary?: string;
  talkingPoints?: string[];
  referralRoutes?: string[];
  recommendedApproach?: string;
}