# lib/api-zod/src/generated/types/sourceOfWealthDraftInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileNoteCoverageItem } from './fileNoteCoverageItem';
import type { SourceOfWealthBriefingInput } from './sourceOfWealthBriefingInput';

export interface SourceOfWealthDraftInput {
  /** @minLength 1 */
  clientName: string;
  /** The banker's meeting note (may be empty when only a briefing is available) */
  note?: string;
  /** Optional confirmed discussion topics to ground the draft */
  coverage?: FileNoteCoverageItem[];
  briefing?: SourceOfWealthBriefingInput;
}