# lib/api-zod/src/generated/types/sourceOfWealthStatement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SourceOfWealthStatement {
  overview: string;
  employment: string;
  compensation: string;
  assetGrowth: string;
  wealthEvents: string;
  plausibility: string;
}